6/A3/requirements.txt
coloredlogs~=15.0
zeroc.ice~=3.7.5